Enable eviction of unresponsive processes and bump version back up to 4.0.4 (#65)
* Evict unresponsive processes and bump NuoDB version to 4.0.4

- This change enables eviction of unresponsive engine processes by
setting the processLivenessCheckSec property to 30 seconds in the admin
values file. This also factors out the pendingReconnectTimeout setting,
which was hardcoded in the template.
- With NuoDB version 4.0.4, the leaderAssignmentTimeout which causes
processes to block while waiting for all archive histories to be
collected and for the initial SM to start up, has been bumped from 10
seconds to 60 seconds. This makes it more likely that the database
teardown in tests will try to delete a database while there are database
processes in this state. This change fixes this by issuing a "shutdown
database" command followed by "check database" command to wait for all
database processes to exit.
- This change bumps the NuoDB version to 4.0.4.