Allow TEs to be defined as a StatefulSet
Add databases.te.enableStatefulSet option that allows TEs to be defined
as a StatefulSet. Assuming the databases.te.dbServices.enabled is not
false, TEs will acquire FQDNs, which are advertised by the Admin and
also appear in the CN attribute of the TE certificate, allowing hostname
verification to be enabled in SQL clients.