Autoconfigure role and service account
Add nuodb.serviceAccount parameter to values files, and automatically
create a service account, role, and role-binding that allows NuoDB Pods
to inspect Kubernetes state.

This is needed by the Admin so that it can clean up objects in the NuoDB
domain state that are associated with resources that are known to be
defunct in Kubernetes. For example, remove a database process object
(start ID) associated with a non-running Pod or container, remove an
archive object (archive ID) associated with a non-existent PVC.

This change also exposes POD_UID as an environment variable to NuoDB
containers.