Add test exercising domain resync (#87)
Add test that exercises unreleased domain resync functionality,
specifically SM scale-down and PVC deletion. For now, this test can only
be run internally, and it also starts multiple SMs, which means that it
requires an Enterprise license. We do not want to publish an Enterprise
license in GitHub, so these tests might remain internal only even after
the domain resync functionality is released.