Remove memoryOption and replace it with resources.requests.memory (#106)
Before NuoDB 4.0.1, it was not possible to use binary prefix number formats (Gi, Mi, Ki) and as such, the memory options had to be split into requests.memory and memoryOption.

Now it is possible to use only one variable to define the memory requirements of an engine (SM, TE). It accepts both scientific/metric units (G, M, K) and binary prefix numbers.

Both formats accept floating type numbers and are not restricted to integers.