add sm.hotCopy.enableBackups to the database helm chart
the sm.hotCopy.enableBackups paramter is used to control whether or
not to enable database backups for the SM hotcopy pods, the default
value is true.

there are use cases where taking a database backup would be better
served if it got trigged by a workflow outside the helm charts
(e.g. on demande backup); it would be better in this case to disable
the kubernetes job/cronjob in the database helm chart that takes a
full backup and periodic incremental backups of a database

Signed-off-by: Abdallah Chatila <abdallah@kaloom.com>