fix TLS rotation test (#121)
The issue in the test was a race condition. A pod can be running, but the admin process is not necessarily ready for requests.

I also added a guard against a race between the restart and the next steps. This race can not be seen in the error above but can be potentially hit.