fix backup job (#123)
The initial backup job can sometimes fail if the DB is not fully initialized when the job is being run.

I decided to change the restart policy to prevent Kubernetes from scheduling multiple pods and possibly accumulating failed jobs. The OnFailure policy makes more sense.

This PR fixes the flaky tests too.

We will have to also re-evaluate the restart policies for our CronJobs. But I am not going to do that in the PR.