fix killSM flaky test (#124)
The issue with the test is two-fold:
A) there can be reasons other than the test itself that cause the SM/TE pods to restart. Assuming that there are 0 restarts is not correct. We should verify that the restart count is higher than it was before the kill.
B) ping-timeout kills the SM during the TE test and hence increases the pod restart count to 1+.

There are three changes in this PR:

add AwaitPodUp, which is not strictly required, but it makes the restart logic easier to reason about
address issue A) by making sure that we first check the number of existing restarts before killing the SM
turn off ping-timeout for the test. This makes the test a bit faster and less random.
Either fix 2 or 3 are required and work stand-alone. I decided to include both.