[DB-30945] Add bootstrapServers label to Admin StatefulSet (#126)
Add boostrapServers label for the Admin StatefulSet that defines the number of
servers that should be part of the initial membership of the domain, which will
be generated by nuodocker start admin. This enables reprovisioning of a fresh
admin-0 instance so it rejoins its peers. Previously, a reprovisioned admin-0
would instance consider itself the lone member of a new domain (since there is
no Raft state to replay, which is the only place where added peers would be
made durable) and would bootstrap a new disjoint domain.

In multi-cluster deployments, this label is only added to the Admin StatefulSet
in the entrypoint cluster.