Remove https in api-server references and add certificates for upgrade tests (#155)
Starting with NuoDB 4.2 the default TLS keys won't be shipped with the image. Users should either generate the keys and expose them as secrets or disable SSL. In case, the user didn't provide overwrite for SSL and the keystore files in the default locations are missing, the SSL will be downgraded to "false". 
This change removes any "https" reference for NUOCMD_API_SERVER variables and adds certificates for upgrade tests.