[HELM-191] Use single-server readiness probe (#166)
Previously, the readiness probe for Admin Pods checked for full-server
connectivity and that all Admin processes agree on a Raft leader. Since these
are not properties of a particular Admin process, but of the entire domain,
they were not appropriate checks for a readiness probe (e.g. admin-1 could show
as not ready because admin-2 is having connectivity issues).

This change adds a readinessprobe script as a ConfigMap that uses the new
'nuocmd check server' (singular) subcommand that will be available in 4.1.2 and
4.2, and falls back to the 'nuocmd check servers' (plural) subcommand that was
previously used as the readiness probe if the new subcommand is not supported.