Remove database name check after restore (#178)
There is a valid use case when a backupset/archive is taken from database with name A and is used to bootstrap database with name B in the same domain. Currently, it's not possible to use autoImport to fulfill the above requirements.
Removed the database name check after the restore is performed.

There shouldn't be a problem with autoRestore because if a backup from a different database is used, the SM startup will fail and complain about a mismatch in the database UUIDs.