Database in-place restore support for NuoDB 4.2+ (#184)
The restore chart has been updated so that it supports `nuodocker request restore`. Support for the new `nuodocker` commands in `nuosm` has been added. The implementation is made so that the "legacy" code is not changed but a new code path is followed in case we have a newer enough product version.