Added stripLevels options in restore chart (#195)
Previously it was not possible to set stripLevels when using the restore chart with remote URL and the default value of 1 was used. Exposed this setting as restore.stripLevels which can be set when installing the restore chart. The value is stored either in the user-data of the database restore request or using $NUODB_RESTORE_REQUEST_PREFIX/$DB_NAME/restore/strip-levels key in legacy mode.