bump minikube (#229)
Newer minikube using driver=none has strict validation of CPU and memory. The CircleCI machine only has 2 CPUs and 7.5GB of RAM. Minikube will also complain if there is not sufficient memory left on the host. It seems that it wants at least 1GB for the host. This leaves us with 6-6.5 useable GiBs.