Add admin helm upgrade workaround to upgrade tests (#239)
When the admin pod gets recreated, the engines do not always reconnect. This is worse in newer versions of K8s.

This is a known issue: https://github.com/nuodb/nuodb-helm-charts/issues/238

Re-enable the tests and add a workaround:
If the engine does not reconnect in time, kill the engine process. This is better than deleting the deployment pod since it won't give the pod a new SHA. This causes at least 2 restarts. 1 restart after the process exits and another restart when the new process gets evicted. There can be as many as 4 restarts before the system stabilizes.

Also refactor the UpgradeOptions as proposed in an earlier PR.