Increase leaderAssignmentTimeout to account for CrashLoopBackOff (#242)
If SMs fail to start due to leaderAssignmentTimeout because some SM is not
being scheduled (e.g. because the volume the archive is on is unavailable),
then eventually the SMs will enter CrashLoopBackOff state and Kubernetes will
begin scheduling them with a 5-minute delay. This would cause the
leaderAssignmentTimeout problems to persist even after all SMs become
schedulable, because now the SMs are more staggered in when they get scheduled,
and we will be lucky for all of them to be scheduled within the same 60-second
window as required by the current leaderAssignmentTimeout setting of 60000
milliseconds.

This change bumps the leaderAssignmentTimeout to 300000 milliseconds (5
minutes) to account for the maximum scheduling delay in Kubernetes.