Update process readiness probe to use information in argv (#252)
Use information about the process that is written into the argument list rather
than relying on "nuocmd show domain" and "nuocmd check process" invocations.
This makes health checks more scalable as the number of databases and processes
in the domain increases. If the process information is not available in argv,
fall back to the old method.