Request incremental hot copy if journal hot copy is temporary disabled (#262)
NuoDB 4.3+ will temporary suspend the journal hot copy during archive sync.
When coordinated hot copy operation is performed, the error is shown either in
the "message" attribute of the hot copy response or when the hot copy is
initiated depending on the choice of the coordinator SM

For example:
- "message": "Hot copy failed: Node 6: Journal hot copy has been suspended due
  to archive sync. Full or incremental hot copy must be completed before
 journal hot copy."
- 'backup database' failed: Failure while performing hot-copy: Error while
  sending engine request: Journal hot copy has been suspended due to archive
  sync. Full or incremental hot copy must be completed before journal hot copy.

If full or incremental hot copy is already requested, the error message will be
different: "Journal hot copy has been suspended due to archive sync. The
running incremental hot copy must complete before journal hot copy can start."

A new incremental hot copy is requested now if the journal hot copy fails
because it is temporary disabled.