Fix TestKubernetesRestoreMultipleSMs test (#270)
The TestKubernetesRestoreMultipleBackupGroups test is doing multiple restore
operations in a sequence which increases the possibility for a container to be
reported as `CrashLoopBackOff.` This increases the testing time due to container
restart back-off and can fail the test. Restart database pods manually after
the restore has been requested.  Database processes that are not selected for
restore may experience schedule delay and start after the restore coordinator
SM already restored the database.  In that case "Waiting for database restore
to complete" message won't be seen in their log.