Update image used by CircleCI (#284)
This change updates the Ubuntu image used to run tests using CircleCI,
updates the version of Minikube used, and removes configuration file
wrangling that is superfluous now that `minikube start` is invoked as
the non-root user.

This change also installs libncurses5, which is required by nuosql and
is not installed by default in newer versions of Ubuntu.