Support for multi-tenancy in a single namespace (#285)
This change enables the installation of multiple domains and databases in a
single Kubernetes namespace. Previously it was not possible to have two admin
releases or two database primary releases with the same database name
collocated in a single Kubernetes namespace. To prevent conflicts regarding
resource naming, the names of existing resources have been changed (see the
table below for details).  Added needed plumbing so that user-defined resource
labels can be injected into all provisioned Kubernetes resources. This can be
used for easy selection of resources that are part of a specific domain or
database.