Allow database.env.NUODB_ARCHIVEDIR and NUODB_JOURNALDIR to be set (#293)
Allow database.env.NUODB_ARCHIVEDIR and NUODB_JOURNALDIR to be set to
override the default of /var/opt/nuodb/archive/<domain>/<dbname>. This
was original intend and then broken when support for restore was added
to nuosm. This change will break any upgrade that currently sets these
variables. I don't believe that case exists,  if so then the variable
settings will need to modified when upgrading.

Close #293