Check if downloaded tarbar is actually an archive (#299)
This patch, checks if downloaded tarbar is actually an archive instead
of backupset. If so then it just renames the directory as if it streamed
the data. So no need for source_type although it is there for backward
compatibility. This code might be better in nuodocker.py instead of the
copy_tree that is done today. But, it is easier to change here.

This is another change that I'm making for Medidata. performance of
stream / backupset for sftp. Was / is based upon nuodocker.py doing a
copy_tree for nonbackupsets when origin-dir and restore-dir are
different. This avoids that by always having origin-dir and restore-dir
being the same for archive restores. By using backupset and checking if
tar uncompressed is actually an archive instead and if so renaming the
download directory to DB_DIR. Since they are both off of DB_PARENTDIR
(same volume) the rename is quick just inode change for top directory.

Close #299