Update TestAdminScaleDown to delete PVC before checking for Raft leader
Deletion of the PVC for the scaled down APs will become a requirement
for future versions of the product. This change updates
TestAdminScaleDown so that it works with the new behavior, and also the
existing behavior, which applies the consensus override immediately
after the pod is deleted after being scaled down.