Bump Kubernetes version for testing (#316)
Changes

- Bump minikube version to 1.30.1
- Bump Kubernetes version to 1.24.12
- Change minikube driver to docker.
- Update the NuoDB image used by the rolling upgrade tests to 4.2.

The `none` driver doesn't seem to work with the newer Kubernetes version due to
failures of the etcd container for an unknown reason. Also, there are several
new dependencies that need to be built and installed which complicates and
delays the test setup process. Using the recommended docker driver simplifies
the setup and doesn't complicate future maintenance.

Kubernetes v1.25 or above is not used because the batch/v1beta1.CronJob API has
been removed. Helm charts v3.4.0 uses this API and we would like to maintain
the upgrade testing coverage from this version.