Support readOnlyRootFilesystem and generic ephemeral volumes
This change adds support for the security feature
readOnlyRootFilesystem, which prevents any modifications from being made
to the root filesystem, and also adds support for generic ephemeral
volumes, which is an alternative to emptyDir that provisions an actual
PVC that is bound to the lifecycle of the pod.

Generic ephemeral volumes can be used in place of emptyDir volumes.
Regardless of what sort of ephemeral storage used for the pod, this
change also volume mounts a subpath of the ephemeral storage into /tmp,
which is a location that NuoDB processes write to that we were not
mounting a volume to.

This change also bumps the default image version to 5.0.3, which
supports readOnlyRootFilesystem. readOnlyRootFilesystem=true for the
admin requires the NUODOCKER_CONF_DIR environment variable to be used to
specify the directory to create the generated nuoadmin.conf file into.
This functionality is available in >=4.3.3 and >=4.2.5.