Failure to record metadata for backups (#334)
Issue

Once the backup index reaches NUODB_MAX_BACKUP_HISTORY, it will be reset to
0. This will cause problems with recording future backup metadata history as
the values in the KV store for these keys are expected to be empty.

Changes

- Set the expected value for the next backup by first fetching the existing
value. This solves the issue at hand while preserving the guard for
concurrent modifications of the backup metadata at that index.
- Make sure that backup indexes always start with 0. Previously the initial
value was 1 but after reaching NUODB_MAX_BACKUP_HISTORY, the index was
reset to 0. Changed the report-backups command to account for that.
- Render database.env Helm values in CronJob allowing users to pass opaque
environment variables.