Add backup hooks sidecar and validate data sources
This change adds a sidecar that exposes HTTP endpoints for pre- and
post-backup hooks to be invoked before and after snapshotting the
archive and journal volumes. This is implemented using the new
backup_hooks.py script, which is mounted into an official Python 3 image
as a configmap.

The backup_hooks.py script controls the lifecycle of a backup by
creating the backup.txt file containing the supplied backup ID when the
`POST /pre-backup/{backupId}` endpoint is invoked. This is used by the
`nuosm` entrypoint script to detect that an archive and journal are
being started for the first time from snapshots.

If the journal volume is separate from the archive volume, then
coordination has to be performed to make sure that the snapshots of the
archive and journal volume are consistent with each other. By default,
this is done by invoking `fsfreeze --freeze` on the archive volume. To
enable testing and to support environments where privileged containers
are not allowed, it is also possible to obtain consistent snapshots by
suspending the nuodb process while snapshots are being taken, which is
enabled by the `database.backupHooks.useSuspend` Helm value.

The `POST /post-backup/{backupId}` endpoint unfreezes write to the
archive or resumes the nuodb process, and deletes the backup.txt file.
It is also possible to directly invoke the backup hooks by using
`kubectl exec` instead of issuing HTTP requests.

This commit also makes a couple of changes to the way data sources are
specified to reduce the potential for misconfiguration. The backup ID is
now nested within `database.snapshotRestore`, which also includes
`snapshotNameTemplate`, which allows the user to define a naming
convention that is used to resolve the snapshots to use as data sources.

The archive and journal data sources can still be specified explicitly
to override the data sources resolved from
`database.snapshotRestore.snapshotNameTemplate`, but it is no longer
possible to supply the hotcopy and non-hotcopy SMs with different data
sources, since that makes no sense (both SM statefulsets use the same
archive data source, so if the journal data sources are different, at
least one of them has to be invalid).

Finally, validation is now performed by default on the resolved data
sources when installing or upgrading a Helm release, to make sure that
the data sources exist. This is necessary because Kubernetes silently
ignores non-existent data sources and allows the volumes to be
initialized with no data, which will lead to confusing behavior.