Start a database from a volume and archive snapshot (#337)
To enable restore and clone use cases, we need a way to start a database from a volume snapshot.

An archive snapshot is specified as a [TypedObjectReference](https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.27/#typedobjectreference-v1-core) and is passed as the helm value `database.persistence.dataSourceRef`. If the snapshot is of an archive without a journal, the journal snapshot should be passed in the helm values `database.sm.hotCopy.journalPath.persistence.dataSourceRef` and `database.sm.noHotCopy.journalPath.persistence.dataSourceRef`.

If the name of the domain or database has changed, there is code to relocate the file and register the moved archive with the admin service.

Unfortunately, this logic cannot differentiate between a snapshot recovered without changing the database and domain names and an SM that was restarted. In order to trigger the registration logic in this case, when taking the snapshot, there needs to be a pre-snapshot script that will create a `backup.txt` file in the archive (and journal, if it is on a separate volume). That file should just contain the backup id as plain text. That same backup id will need to be set as the helm value `database.autoImport.backup_id`. At restore time, the file will be deleted to prevent the archive from being registered twice.