[HELM-411] Multi-cluster test configuration migration from Azure to AWS
The names of the clusters have been changed and required changes in config.yaml.
A section is added of AWS is added in install_deps.sh to update the kubeconfig.
The EKS cluster names and AWS credentials are added as ENV on CircleCi.
The helm chart release name in 'StartDatabaseTemplate' in nuodb_database_utilities.go
has been truncated as the pod name has a 64 chars limit.