Update DNS configuration before multi-cluster tests
The IPs for the elastic load balancer endpoints that are used to expose
each DNS server are not permanent, so the CoreDNS configuration has to
be updated to include the current IPs. This change programmatically
resolves the IPs of each DNS load balancer and adds them as upstream
resolvers for the other.

The "Corefile" configmap entry is assumed to contain the marker string
"# BEGIN BASE CONFIG", which delimits the non-changing portion of the
config and enables updates to the generated portion.