Account for gap between snapshot preparation and archive creation (#361)
For SMs created from snapshots, we currently we rely on the absence of
the info.json file and presence of the backup.txt file to detect that
the archive has to be prepared from a snapshot. But there is a window of
time until the info.json file for the newly-prepared archive is created
in which any failure would leave the `nuosm` script unable to detect if
it should perform a normal restart.

This change adds a restore.txt file which is created at the same time
the old info.json and backup.txt file are deleted and is used to detect
that the SM already performed snapshot preparation and should continue
with normal startup.