Tighten semver parsing for nuosql version (#363)
This change tightens the parsing of the nuosql version to account for
any non-numerical suffix that is not delimited by a dash (`-`). For
example, development versions on release branches have the form
`<major>.<minor>.<patch>.<comment>-<build num>-<SHA>-<arch>`. The
`<comment>` part of the version was not being stripped away and screwed
up the semver parsing.