Support for HotSnap in backup hooks (#365)
This change allows using `hotsnap` (available in NuoDB 6.0.2 and above) to
pause/resume SM archiving in the backup hooks container. A new Helm value
`database.backupHooks.freezeMode` is added so that the user can select the mode
in which the backup hook service operates.

Added `database.backupHooks.timeout` (30s by default) which defines the timeout
in seconds after which the backup is automatically canceled if the post-backup
hook is not invoked. This is generally useful since it guards against
situations where the post-backup hook is not called to unfreeze the archive
volume. Automatic backup canceling is mandatory when "hotsnap" mode is
activated. The SM automatically resumes the archiving and the backup hook
server must remove the backup metadata file.

When "hotsnap" mode is used, an in-progress backup is invalidated on SM
restart. The backup hook server stores the SM start ID and the backup ID when
the pre-backup hook is called. This allows it to detect SM restart and remove
the backup metadata in such cases.

Deprecated the `database.backupHooks.useSuspend` Helm value which is replaced
by `database.backupHooks.freezeMode=suspend`.