Fixed two issues in TestKubernetesUpgradeFullDatabase (#374)
`TestKubernetesUpgradeFullDatabase/verifyProtocolVersion` was failing on NuoDB 5.1, because there is no protocol upgrade between 5.0 and 5.1. Change the test to only run on 6.0 and newer.

While I was investigating the issue, I found a race condition in the test. The the protocol upgrade is scaling up the TE count to 2 and upgrading the database version in a single Helm update. It assumes that if there the database has 2 TEs and is healthy, the upgrade is done, but it can also occur part way through the upgrade if one of the new TEs becomes running before the admin sees the second new TE.