Wait longer for AP readiness in nuosm
This change increases the amount of time to wait for the APs to elect a
Raft leader and changes the retry function to use an exponentially
increasing polling interval until a timeout expires, rather than using a
fixed interval of 1 second for a specific number of iterations.

The reason for this change is to prevent nuosm from failing fast,
generating many unnecessary restarts, and causing the SM container to
enter CrashLoopBackOff, which delays database creation. Using an
exponentially increasing interval (capped at 10 seconds) also reduces
the amount of logging noise.