Add test of nuodocker functionality to delete old archives
This change adds a test of `nuodocker start sm` functionality that
detects and deletes the old archive object associated with a PVC. This
functionality will be delivered in 7.0 and perhaps backported to 6.0.3.

If the NuoDB version does not support this functionality, then the test
asserts on the expected current behavior, which is that database restart
is blocked waiting for an SM to be started on the defunct archive.