Allow HTTP requests to be dispatched to scripts
This change adds the ability to register custom handlers for the HTTP
server in the backup hooks sidecar container. Each handler has a method,
path, and script, which is executed on incoming reqeusts with matching
method and path. Any path or query parameters extracted from the URL can
be referenced as environment variables in the script, as well as the
request payload.