Support static NuoDB collector configuration (#400)
This change enables the NuoDB collector to be configured with plugins
statically by mounting ConfigMap resources directly in the container. It is
useful for deployments where it is not needed to pull the Telegraf plugins
dynamically (i.e. all plugins are known at the Helm release installation time).
The immediate benefit of such a configuration is disabling the NuoDB collector
config watcher sidecar, to reduce the overall cost at scale and minimize
the load on the Kubernetes API server.

The plugin configuration changes are automatically propagated inside the
container by kubelet, however, Telegraf must be started with `--watch-config`
option. See https://docs.influxdata.com/telegraf/v1/commands/