Allow injection of scripts to enable data migration
This change adds the ability to inject custom scripts for performing
data migration.

If data migration is enabled, the supplied `prepare-archive` script is
invoked early in the `nuosm` entrypoint script in order to prepare the
archive directory so that it gets imported into the new database.

The `load-credentials` is sourced by the readiness probe for the SM with
ordinal 0 in order to obtain the DBA credentials needed in order to
update the DBA user and password so that it matches the configuration
supplied for the new database. The SM with ordinal is not marked as
ready until it performs this update.